requirements.txt
# requirements.txt

streamlit
pandas
python-docx